Split the files so that it's more digestible
