+++ app/src/main/kotlin/org/db_poultry/App.kt
@@ -5,6 +5,7 @@ import javafx.application.Application
 import org.db_poultry.controller.MainFrame
 import org.db_poultry.db.DBConnect
 import org.db_poultry.errors.generateErrorMessage
+import org.db_poultry.theLifesaver.Backup.TL_checkLastBackupDate
 import org.db_poultry.theLifesaver.TL.*
 import java.sql.Connection
 
@@ -90,5 +91,5 @@ fun main() {
     // ==================================================
     // Keep this here but remove before shipping or every release
     // ==================================================
-    TL_cleanAll()
+    wipe()
 }
